Color 66083's checklists.
These colors match 73146's checklists. I changed a few of 146's colors to avoid putting two checklists of identical color next to each other in 083's checklist. I also swapped the colors of the "Engine Failure During Takeoff Run" and "Ditching" checklists because I think it makes sense for the Ditching box to be light blue.

I also bumped the brightness of the purple up a bit to make it easier to distinguish from brown in slightly-low lighting.