Produce print-shop PDF with every even page rotated.
The print shop's printers can't do double-sided printing on 100 lb paper. Manually flipping the page results in both sides being misaligned. This change produces PDF files with every other page rotated, which allows us to manually perform double-sided printing without having the sides be misaligned.

The rotation is done using the qpdf tool. qpdf simply updates the rotation angle; it does not re-encode the PDF, so it should not result in quality loss. Note that some PDF readers will not show it correctly, but I've already tested that the print shop printers handle it well.

To make this all automatic, I added a `Makefile`. To simplify the CI environment, I also added support for generating svgs to the Makefile, and adapted the workflow files accordingly.