Tweak colors further.
I just got back from the print shop with updated samples. This tweaks the colors a bit more. Adjusting by only 1 as I did before to avoid overshooting.

The one exception is grey: in PR #59 I accidentally tweaked it in the wrong direction (darker when I intended lighter). This corrects it to the value I intended to set in that PR.