More fabrication updates based on today's experiments.
I had success with this technique as-written, with the sole exception that I did not bring a book to set on the lamination while it cools. In that past that was never an issue (the folding process took long enough that it was cool by the time I wanted to lay my weight on it). This time, it was still too hot while I was pressing my hand down (ouch).