Preflight tweaks: oil order, ARROW docs.
This moves the oil check from the "Fluids" section to the "Nose" section, to make it more consistent with the POHs and to better fit it into the preflight flow. The "Fluids" section was renamed to "Fuel" because now that's all it includes (other than the ignition check, which I'm keeping up front).

I also changed the step about verifying the POH is accessible with one to verify all the ARROW docs are good.