Copy 73146.typ to 73063.typ, update post-comment action.
This PR copies 73146.typ to 73063.type, but does not make any changes. I will make the changes in a separate PR so that the changes are clear during code review. This does proactively change the post-comment action, so that when that later PR is sent through 063's checklist is rendered.