Finish 73063's checklist.
I went back through the 1976 172M POH as well as the documentation for 73063's avionics to update the checklist for it. This also adjusts all the other files in the repository that need to be updated (Makefile, README, and .gitignore).

This is written assuming that 063 does NOT have strobes. If it does have strobes, this will need to be updated (it shows up in a few normal operating checklists as well as the wing fire checklist).